remove unnecessary conversion

+++ storagenode/pieces/readwrite.go
@@ -79,7 +79,7 @@ func NewReader(blob storage.BlobReader, bufferSize int) (*Reader, error) {
 	reader := &Reader{}
 	reader.buf = *bufio.NewReaderSize(blob, bufferSize)
 	reader.blob = blob
-	reader.size = int64(size)
+	reader.size = size
 
 	return reader, nil
 }
